+++ app/Http/Controllers/TagController.php
@@ -4,9 +4,15 @@ namespace App\Http\Controllers;
 
 use App\Tag;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;  // allow functions only through policies
 
 class TagController extends Controller
 {
+    public function __construct() 
+    {
+      $this->middleware('auth')->except(['index']); // require auth everywhere but for index and show views
+      $this->middleware('admin')->except(['index']);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -64,7 +70,9 @@ class TagController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function edit(Tag $tag)
-    {        
+    {   
+      abort_unless(Gate::allows('update', $tag), 403);  // gate policy 
+           
       return view('tag.edit')->with([
         'tag'=>$tag
       ]);
@@ -79,6 +87,7 @@ class TagController extends Controller
      */
     public function update(Request $request, Tag $tag)
     {
+      abort_unless(Gate::allows('update', $tag), 403);  // gate policy 
       $request->validate([
         'name' => 'required', 
         'style' => 'required'
@@ -104,6 +113,8 @@ class TagController extends Controller
      */
     public function destroy(Tag $tag)
     {
+      abort_unless(Gate::allows('delete', $tag), 403);  // gate policy
+
       $oldName = $tag->name;
       $tag->delete();
